requirements.txt
fastapi>=0.85.0
uvicorn\[standard]>=0.18.0
langchain>=0.1.XX
requests>=2.28.0
pillow>=9.2.0
rembg>=2.0.0
torch>=1.12.0
diffusers>=0.16.0
